WIP: Move refresh action from a text into a button in user view
Brings top navigation bar to user view.

Seems that there is a bug that I didn't figure a cause for:
When moving from another main view to user view, and pressing settings
button to reveal language choise, then moving back to another view and
finally back to user view, which was left to settings view. Now, pressing
back does not bring the user view, but again settings view. This get's
fixed when moving to another view and back to user view.