Make swipe transition speedier (by using onWillFocus)
Use onWillFocus and update Redux state before calling navigator.pop().
This way, when the transition starts, we can rely on redux state.

The idea is that onWillFocus is called both on swipe and on back button
press (navigator.pop()). In the former case, we detect that redux state
needs to be updated. In the latter case, redux state has already been
updated so we don't need to do anything. For this to work, redux state
has to change before onWillFocus starts, which is triggered by
navigator.pop().